Move p2p keypair initialization to p2p mod
